fixes #1
SpringLint was throwing an exception when finding an annotation declaration. That was due to the fact that the ClassVisitor wasn't ready for @interface. Now, it can extract the name of this kind of type.

As an addition, the smell detector now ignores anything that is not a class (e.g. interface, annotation). That should improve performance a bit.